LayoutTests/http/tests/contentextensions/text-track-blocked-expected.txt

CONSOLE MESSAGE: video loaded
CONSOLE MESSAGE: Content blocker prevented frame displaying http://127.0.0.1:8000/contentextensions/text-track-blocked.html from loading a resource from http://127.0.0.1:8000/media-resources/track/captions-webvtt/captions-fast.vtt
CONSOLE MESSAGE: vttError
The following text track should not load.
